Fix fluid clearing causing solar "brownouts"
Instead of clearing all solar fluid in a tower and adding it back, we only add or remove fluid as needed to bring it to the target amount.

This fixes laser power towers turning off for a single tick when refilling their solar fluid, and from my testing has no other side effects on either tower type.